refactor: split preview runtime chunks
(cherry picked from commit 4b4d994db94c21a1ad83198ef92991499a84f559)
